yuan,feat: 状态模式实现活动状态之间的流转，依赖倒置基础设施层引用领域层

+++ lottery-interfaces/src/test/java/com/yuanstack/lottery/infrastructure/dao/ActivityDaoTest.java
@@ -34,7 +34,7 @@ public class ActivityDaoTest {
         activity.setEndDateTime(new Date());
         activity.setStockCount(100);
         activity.setTakeCount(10);
-        activity.setState((byte) 0);
+        activity.setState(0);
         activity.setCreator("hhhh");
         activityDao.insert(activity);
     }
